requirements.txt
Pillow
nvidia-ml-py
PyQt5
pyinstaller
